refactor: extracted PositionNodeFinder into _position_node_finder.py
